+++ R/cv.enspls.R
@@ -21,7 +21,7 @@
 #'
 #' @note To maximize the probablity that each observation can
 #' be selected in the test set (thus the prediction uncertainty
-#' can be measured), please try setting a large \code{MCtimes}.
+#' can be measured), please try setting a large \code{reptimes}.
 #'
 #' @seealso See \code{\link{enspls.fit}} for ensemble sparse
 #' partial least squares regression.
@@ -36,7 +36,7 @@
 #' y = logd1k$y
 #'
 #' set.seed(42)
-#' cvfit = cv.enspls(x, y, MCtimes = 10)
+#' cvfit = cv.enspls(x, y, reptimes = 10)
 #' print(cvfit)
 #' plot(cvfit)}