+++ R/enpls.fs.R
@@ -6,7 +6,8 @@
 #' @param y response vector
 #' @param maxcomp Maximum number of components included within the models,
 #' if not specified, default is the variable (column) numbers in x.
-#' @param MCtimes times of Monte-Carlo
+#' @param reptimes Number of models to build with Monte-Carlo resampling
+#' or bootstrapping.
 #' @param method \code{"mc"} or \code{"bootstrap"}. Default is \code{"mc"}.
 #' @param ratio sample ratio used when \code{method = "mc"}
 #' @param parallel Integer. Number of CPU cores to use.
@@ -29,25 +30,19 @@
 #' @importFrom doParallel registerDoParallel
 #' @importFrom foreach foreach "%dopar%"
 #'
-#' @references
-#' Dongsheng Cao, Yizeng Liang, Qingsong Xu, Yifeng Yun, and Hongdong Li.
-#' "Toward better QSAR/QSPR modeling: simultaneous outlier detection and
-#' variable selection using distribution of model features."
-#' \emph{Journal of computer-aided molecular design} 25, no. 1 (2011): 67--80.
-#'
 #' @examples
 #' data("alkanes")
 #' x = alkanes$x
 #' y = alkanes$y
 #'
 #' set.seed(42)
-#' fs = enpls.fs(x, y, MCtimes = 50)
+#' fs = enpls.fs(x, y, reptimes = 50)
 #' print(fs)
 #' plot(fs)
 
 enpls.fs = function(x, y,
                     maxcomp = NULL,
-                    MCtimes = 500L,
+                    reptimes = 500L,
                     method = c('mc', 'bootstrap'), ratio = 0.8,
                     parallel = 1L) {
 
@@ -58,20 +53,20 @@ enpls.fs = function(x, y,
   method = match.arg(method)
 
   x.row = nrow(x)
-  samp.idx = vector('list', MCtimes)
+  samp.idx = vector('list', reptimes)
 
   if (method == 'mc') {
-    for (i in 1L:MCtimes) samp.idx[[i]] = sample(1L:x.row, round(x.row * ratio))
+    for (i in 1L:reptimes) samp.idx[[i]] = sample(1L:x.row, round(x.row * ratio))
   }
 
   if (method == 'bootstrap') {
-    for (i in 1L:MCtimes) samp.idx[[i]] = sample(1L:x.row, x.row, replace = TRUE)
+    for (i in 1L:reptimes) samp.idx[[i]] = sample(1L:x.row, x.row, replace = TRUE)
   }
 
   if (parallel < 1.5) {
 
-    coeflist = vector('list', MCtimes)
-    for (i in 1L:MCtimes) {
+    coeflist = vector('list', reptimes)
+    for (i in 1L:reptimes) {
       xtmp = x[samp.idx[[i]], ]
       xtmp = scale(xtmp, center = TRUE, scale = TRUE)
       ytmp = y[samp.idx[[i]]]
@@ -82,7 +77,7 @@ enpls.fs = function(x, y,
   } else {
 
     registerDoParallel(parallel)
-    coeflist = foreach(i = 1L:MCtimes) %dopar% {
+    coeflist = foreach(i = 1L:reptimes) %dopar% {
       xtmp = x[samp.idx[[i]], ]
       xtmp = scale(xtmp, center = TRUE, scale = TRUE)
       ytmp = y[samp.idx[[i]]]