+++ R/enpls.od.R
@@ -6,7 +6,8 @@
 #' @param y response vector
 #' @param maxcomp Maximum number of components included within the models,
 #' if not specified, default is the variable (column) numbers in x.
-#' @param MCtimes times of Monte-Carlo
+#' @param reptimes Number of models to build with Monte-Carlo resampling
+#' or bootstrapping.
 #' @param method \code{"mc"} or \code{"bootstrap"}. Default is \code{"mc"}.
 #' @param ratio sample ratio used when \code{method = "mc"}
 #' @param parallel Integer. Number of CPU cores to use.
@@ -24,7 +25,7 @@
 #'
 #' @note To maximize the probablity that each observation can
 #' be selected in the test set (thus the prediction uncertainty
-#' can be measured), please try setting a large \code{MCtimes}.
+#' can be measured), please try setting a large \code{reptimes}.
 #'
 #' @seealso See \code{\link{enpls.fs}} for feature selection with
 #' ensemble partial least squares regression.
@@ -35,30 +36,20 @@
 #' @importFrom doParallel registerDoParallel
 #' @importFrom foreach foreach "%dopar%"
 #'
-#' @references
-#' DongSheng Cao, Yizeng Liang, Qingsong Xu, Hongdong Li, and Xian Chen.
-#' "A new strategy of outlier detection for QSAR/QSPR."
-#' \emph{Journal of computational chemistry} 31, no. 3 (2010): 592--602.
-#'
-#' Dongsheng Cao, Yizeng Liang, Qingsong Xu, Yifeng Yun, and Hongdong Li.
-#' "Toward better QSAR/QSPR modeling: simultaneous outlier detection and
-#' variable selection using distribution of model features."
-#' \emph{Journal of computer-aided molecular design} 25, no. 1 (2011): 67--80.
-#'
 #' @examples
 #' data("alkanes")
 #' x = alkanes$x
 #' y = alkanes$y
 #'
 #' set.seed(42)
-#' od = enpls.od(x, y, MCtimes = 100)
+#' od = enpls.od(x, y, reptimes = 100)
 #' print(od)
 #' plot(od)
 #' plot(od, criterion = 'sd')
 
 enpls.od = function(x, y,
                     maxcomp = NULL,
-                    MCtimes = 500L,
+                    reptimes = 500L,
                     method = c('mc', 'bootstrap'), ratio = 0.8,
                     parallel = 1L) {
 
@@ -69,18 +60,18 @@ enpls.od = function(x, y,
   method = match.arg(method)
 
   x.row = nrow(x)
-  samp.idx = vector('list', MCtimes)
-  samp.idx.remain = vector('list', MCtimes)
+  samp.idx = vector('list', reptimes)
+  samp.idx.remain = vector('list', reptimes)
 
   if (method == 'mc') {
-    for (i in 1L:MCtimes) {
+    for (i in 1L:reptimes) {
       samp.idx[[i]] = sample(1L:x.row, round(x.row * ratio))
       samp.idx.remain[[i]] = setdiff(1L:x.row, samp.idx[[i]])
     }
   }
 
   if (method == 'bootstrap') {
-    for (i in 1L:MCtimes) {
+    for (i in 1L:reptimes) {
       samp.idx[[i]] = sample(1L:x.row, x.row, replace = TRUE)
       samp.idx.remain[[i]] = setdiff(1L:x.row, unique(samp.idx[[i]]))
     }
@@ -90,8 +81,8 @@ enpls.od = function(x, y,
 
   if (parallel < 1.5) {
 
-    errorlist = vector('list', MCtimes)
-    for (i in 1L:MCtimes) {
+    errorlist = vector('list', reptimes)
+    for (i in 1L:reptimes) {
       plsdf.sample = plsdf[samp.idx[[i]], ]
       plsdf.remain = plsdf[samp.idx.remain[[i]], ]
       errorlist[[i]] = suppressWarnings(enpls.od.core(plsdf.sample, plsdf.remain, maxcomp))
@@ -100,7 +91,7 @@ enpls.od = function(x, y,
   } else {
 
     registerDoParallel(parallel)
-    errorlist = foreach(i = 1L:MCtimes) %dopar% {
+    errorlist = foreach(i = 1L:reptimes) %dopar% {
       plsdf.sample = plsdf[samp.idx[[i]], ]
       plsdf.remain = plsdf[samp.idx.remain[[i]], ]
       enpls.od.core(plsdf.sample, plsdf.remain, maxcomp)
@@ -108,8 +99,8 @@ enpls.od = function(x, y,
 
   }
 
-  prederrmat = matrix(NA, ncol = x.row, nrow = MCtimes)
-  for (i in 1L:MCtimes) {
+  prederrmat = matrix(NA, ncol = x.row, nrow = reptimes)
+  for (i in 1L:reptimes) {
     for (j in 1L:length(samp.idx.remain[[i]])) {
       prederrmat[i, samp.idx.remain[[i]][j]] = errorlist[[i]][j]
     }