+++ R/enspls.fs.R
@@ -8,7 +8,8 @@
 #' if not specified, default is 5.
 #' @param alpha Parameter (grid) controlling sparsity of the model.
 #' If not specified, default is \code{seq(0.2, 0.8, 0.2)}.
-#' @param MCtimes times of Monte-Carlo
+#' @param reptimes Number of models to build with Monte-Carlo resampling
+#' or bootstrapping.
 #' @param method \code{"mc"} or \code{"bootstrap"}. Default is \code{"mc"}.
 #' @param ratio sample ratio used when \code{method = "mc"}
 #' @param parallel Integer. Number of CPU cores to use.
@@ -38,7 +39,7 @@
 #' y = logd1k$y
 #'
 #' set.seed(42)
-#' fs = enspls.fs(x, y, MCtimes = 5, maxcomp = 2)
+#' fs = enspls.fs(x, y, reptimes = 5, maxcomp = 2)
 #' print(fs, nvar = 10)
 #' plot(fs, nvar = 10)
 #' plot(fs, type = 'boxplot', limits = c(0.05, 0.95), nvar = 10)
@@ -46,7 +47,7 @@
 enspls.fs = function(x, y,
                      maxcomp = 5L,
                      alpha = seq(0.2, 0.8, 0.2),
-                     MCtimes = 500L,
+                     reptimes = 500L,
                      method = c('mc', 'bootstrap'), ratio = 0.8,
                      parallel = 1L) {
 
@@ -55,20 +56,20 @@ enspls.fs = function(x, y,
   method = match.arg(method)
 
   x.row = nrow(x)
-  samp.idx = vector('list', MCtimes)
+  samp.idx = vector('list', reptimes)
 
   if (method == 'mc') {
-    for (i in 1L:MCtimes) samp.idx[[i]] = sample(1L:x.row, round(x.row * ratio))
+    for (i in 1L:reptimes) samp.idx[[i]] = sample(1L:x.row, round(x.row * ratio))
   }
 
   if (method == 'bootstrap') {
-    for (i in 1L:MCtimes) samp.idx[[i]] = sample(1L:x.row, x.row, replace = TRUE)
+    for (i in 1L:reptimes) samp.idx[[i]] = sample(1L:x.row, x.row, replace = TRUE)
   }
 
   if (parallel < 1.5) {
 
-    coeflist = vector('list', MCtimes)
-    for (i in 1L:MCtimes) {
+    coeflist = vector('list', reptimes)
+    for (i in 1L:reptimes) {
       xtmp = x[samp.idx[[i]], ]
       xtmp = scale(xtmp, center = TRUE, scale = TRUE)
       ytmp = y[samp.idx[[i]]]
@@ -78,7 +79,7 @@ enspls.fs = function(x, y,
   } else {
 
     registerDoParallel(parallel)
-    coeflist = foreach(i = 1L:MCtimes) %dopar% {
+    coeflist = foreach(i = 1L:reptimes) %dopar% {
       xtmp = x[samp.idx[[i]], ]
       xtmp = scale(xtmp, center = TRUE, scale = TRUE)
       ytmp = y[samp.idx[[i]]]