+++ R/enspls.od.R
@@ -8,7 +8,8 @@
 #' if not specified, default is 5.
 #' @param alpha Parameter (grid) controlling sparsity of the model.
 #' If not specified, default is \code{seq(0.2, 0.8, 0.2)}.
-#' @param MCtimes times of Monte-Carlo
+#' @param reptimes Number of models to build with Monte-Carlo resampling
+#' or bootstrapping.
 #' @param method \code{"mc"} or \code{"bootstrap"}. Default is \code{"mc"}.
 #' @param ratio sample ratio used when \code{method = "mc"}
 #' @param parallel Integer. Number of CPU cores to use.
@@ -26,7 +27,7 @@
 #'
 #' @note To maximize the probablity that each observation can
 #' be selected in the test set (thus the prediction uncertainty
-#' can be measured), please try setting a large \code{MCtimes}.
+#' can be measured), please try setting a large \code{reptimes}.
 #'
 #' @seealso See \code{\link{enspls.fs}} for feature selection with
 #' ensemble sparse partial least squares regression.
@@ -44,14 +45,14 @@
 #' y = logd1k$y
 #'
 #' set.seed(42)
-#' od = enspls.od(x, y, MCtimes = 5, maxcomp = 3)
+#' od = enspls.od(x, y, reptimes = 5, maxcomp = 3)
 #' plot(od, prob = 0.1)
 #' plot(od, criterion = "sd", sdtimes = 1)
 
 enspls.od = function(x, y,
                      maxcomp = 5L,
                      alpha = seq(0.2, 0.8, 0.2),
-                     MCtimes = 500L,
+                     reptimes = 500L,
                      method = c('mc', 'bootstrap'), ratio = 0.8,
                      parallel = 1L) {
 
@@ -60,18 +61,18 @@ enspls.od = function(x, y,
   method = match.arg(method)
 
   x.row = nrow(x)
-  samp.idx = vector('list', MCtimes)
-  samp.idx.remain = vector('list', MCtimes)
+  samp.idx = vector('list', reptimes)
+  samp.idx.remain = vector('list', reptimes)
 
   if (method == 'mc') {
-    for (i in 1L:MCtimes) {
+    for (i in 1L:reptimes) {
       samp.idx[[i]] = sample(1L:x.row, round(x.row * ratio))
       samp.idx.remain[[i]] = setdiff(1L:x.row, samp.idx[[i]])
     }
   }
 
   if (method == 'bootstrap') {
-    for (i in 1L:MCtimes) {
+    for (i in 1L:reptimes) {
       samp.idx[[i]] = sample(1L:x.row, x.row, replace = TRUE)
       samp.idx.remain[[i]] = setdiff(1L:x.row, unique(samp.idx[[i]]))
     }
@@ -79,8 +80,8 @@ enspls.od = function(x, y,
 
   if (parallel < 1.5) {
 
-    errorlist = vector('list', MCtimes)
-    for (i in 1L:MCtimes) {
+    errorlist = vector('list', reptimes)
+    for (i in 1L:reptimes) {
       x.sample = x[samp.idx[[i]], ]
       x.remain = x[samp.idx.remain[[i]], ]
       y.sample = y[samp.idx[[i]]]
@@ -92,7 +93,7 @@ enspls.od = function(x, y,
   } else {
 
     registerDoParallel(parallel)
-    errorlist = foreach(i = 1L:MCtimes) %dopar% {
+    errorlist = foreach(i = 1L:reptimes) %dopar% {
       x.sample = x[samp.idx[[i]], ]
       x.remain = x[samp.idx.remain[[i]], ]
       y.sample = y[samp.idx[[i]]]
@@ -102,8 +103,8 @@ enspls.od = function(x, y,
 
   }
 
-  prederrmat = matrix(NA, ncol = x.row, nrow = MCtimes)
-  for (i in 1L:MCtimes) {
+  prederrmat = matrix(NA, ncol = x.row, nrow = reptimes)
+  for (i in 1L:reptimes) {
     for (j in 1L:length(samp.idx.remain[[i]])) {
       prederrmat[i, samp.idx.remain[[i]][j]] = errorlist[[i]][j]
     }